добавление в data.js функции по созданию данных
